add argument delete meta.json
Deleting the meta.json is useful when you need a metadata refresh, but need to keep existing taken images and such.

Not as intrusive as cleanup which kills the entire tmp directory.

Should be most useful for remote installations where accessing the folder is more painful than a simple argument.