Fix edit-args
During the refactor of the process in V2, one unintended consequence was background looping of the tracker_status process, resulting in tracker confirmation prompts equally the amount of times meta was edited via `is this correct`.

This refactor seeks to unfuck things.

Will also attempt to refactor the edit arg process to properly track original and edited meta. ie: If someone specifies an argument to be modified, that argument should stick, even if the arguments are modified 37 times until the user finally figures it out. *low priority*

Also seek to catch any edge cases where an argument, ie: trackers, does not modify correctly.